WiproAssignments/javaAssignments/javaAssignments.day8.part2/src/main/java/javaAssignments/day8/part2/task4/MathematicalOperations.java
/* Task 1: Write a set of JUnit tests for a given class with simple mathematical operations (add, subtract, multiply, divide) using the basic @Test annotation.
 */

package javaAssignments.day8.part2.task4;

public class MathematicalOperations {

    public int add(int a, int b) {
        return a + b;
    }

    public int subtract(int a, int b) {
        return a - b;
    }

    public int multiply(int a, int b) {
        return a * b;
    }

    public double divide(int dividend, int divisor) {
        if (divisor == 0) {
            throw new IllegalArgumentException("Cannot divide by zero");
        }
        return (double) dividend / divisor;
    }
}

